+++ pybamm/input/parameters/lithium-ion/negative_electrodes/graphite_Ramadass2004/graphite_ocp_Ramadass2004.py
@@ -24,9 +24,3 @@ def graphite_ocp_Ramadass2004(sto):
     )
 
     return u_eq
-
-
-# import pybamm
-
-# sto = pybamm.linspace(0, 1,1000)
-# pybamm.plot(sto, graphite_ocp_Ramadass2004(sto))