+++ pybamm/parameters/lithium_ion_parameters.py
@@ -131,7 +131,7 @@ class LithiumIonParameters:
         }
         self.epsilon_p = pybamm.FunctionParameter("Positive electrode porosity", inputs)
 
-        self.epsilon = pybamm.Concatenation(
+        self.epsilon = pybamm.concatenation(
             self.epsilon_n, self.epsilon_s, self.epsilon_p
         )
         self.epsilon_inactive_n = (
@@ -807,7 +807,7 @@ class LithiumIonParameters:
         self.epsilon_n_init = pybamm.Parameter("Negative electrode porosity")
         self.epsilon_s_init = pybamm.Parameter("Separator porosity")
         self.epsilon_p_init = pybamm.Parameter("Positive electrode porosity")
-        self.epsilon_init = pybamm.Concatenation(
+        self.epsilon_init = pybamm.concatenation(
             self.epsilon_n, self.epsilon_s, self.epsilon_p
         )
         self.T_init = self.therm.T_init